+++ tests/test_integration.py
@@ -1,17 +1,14 @@
-import collections
 import json
-import uuid
 
 from marshmallow_jsonapi import Schema
 from marshmallow_jsonapi import fields
 
 from flask_jsonapi import api
 from flask_jsonapi import resource_repositories
-from flask_jsonapi import nested_resource_repositories
-from flask_jsonapi.nested_repository import ChildRepository
-from flask_jsonapi.marshmallow_nested_extension.schema import IdMappingSchema
 from flask_jsonapi.marshmallow_nested_extension.field import CompleteNestedRelationship
-
+from flask_jsonapi.marshmallow_nested_extension.schema import IdMappingSchema
+from flask_jsonapi.nested import nested_resource_repositories
+from flask_jsonapi.nested.nested_repository import ChildRepository
 
 JSONAPI_HEADERS = {'content-type': 'application/vnd.api+json', 'accept': 'application/vnd.api+json'}
