+++ tests/test_nested_repositories.py
@@ -3,8 +3,8 @@ from unittest import mock
 
 import pytest
 
-from flask_jsonapi import nested_repository
 from flask_jsonapi import resource_repositories
+from flask_jsonapi.nested import nested_repository
 
 
 class Repository:
